big refactoring in order to have paste token functionality in place
