+++ src/pages/Authentication/components/Input/InputForm.tsx
@@ -1,0 +1,31 @@
+import styles from './styles.module.scss';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faPaste } from '@fortawesome/free-solid-svg-icons';
+import { Textarea } from './components/Textarea';
+import { Status } from './components/Status';
+import { Form } from 'formik';
+import { useInputActions } from './hooks/useInputActions';
+
+export const InputForm = () => {
+  const { handleGenerateToken, handlePasteToken } = useInputActions();
+
+  return (
+    <Form className={styles.form}>
+      <h3 className={styles.subtitle}>
+        <span>Paste a token here</span>
+
+        <div className={styles.wrapper}>
+          <button onClick={handlePasteToken} className={styles.paste}>
+            <FontAwesomeIcon icon={faPaste} size='lg' />
+          </button>
+          <button onClick={handleGenerateToken} className={styles.generate}>
+            Generate
+          </button>
+        </div>
+      </h3>
+
+      <Textarea />
+      <Status />
+    </Form>
+  );
+};
