+++ src/pages/Authentication/constants/metrics.contants.ts
@@ -1,0 +1,11 @@
+import { MetricType } from '../types';
+
+export const emptyMetrics: MetricType = {
+  address: '',
+  blockHash: '',
+  signature: '',
+  body: '',
+  host: '',
+  ttl: 0,
+  extraInfo: {}
+};
